Add react-router v4 release to the repository.
Reason for using v4 release is it supports BrowserRouter and ServerRouter to serve JS and non-JS versions.
As this is a longer term project and will not be available in production for a while the alpha release is not an issue.
The hope is that by the time this solution matures the v4 of react-router will be released.